Conflicts:
	‏‏Assignment 2/ZerLi_Client/src/entities/Survey.java
	‏‏Assignment 2/ZerLi_Server/.classpath
	‏‏Assignment 2/ZerLi_Server/src/entities/Survey.java

+++ ‏‏Assignment 2/ZerLi_Client/src/lior/interfaces/IOrderReportController.java
@@ -1,10 +1,9 @@
 package lior.interfaces;
 
+import java.math.BigInteger;
 import java.text.ParseException;
-import java.util.ArrayList;
 import java.util.Date;
 
-import entities.OrderReport;
 import interfaces.IParent;
 
 public interface IOrderReportController extends IParent  {
@@ -15,5 +14,5 @@ public interface IOrderReportController extends IParent  {
 	 * @param storeID
 	 * @throws ParseException 
 	 */
-	void produceOrderReport(Date reqDate, int storeID) throws ParseException;
+	void produceOrderReport(Date reqDate, BigInteger storeID) throws ParseException;
 }
